Add files via upload
Attached is a demo of fitting some scipy.stats distributions to a given set of data (in this case, the duration newspaper subscriptions), and determining which are best from a comparison of Q-Q (quantile-quantile) plots and p values (the likelihood that the given data could be generated from a random distribution with the found parameters). These distributions could then be used to model the data, simulating what can be expected if something about the system were to be changed. This could be used to give client companies a way to "test out" new business strategies without needing to risk their actual business. The entire process can be automated very easily, and random samples can be generated even more easily.